cllm/requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.27.0
faiss-cpu>=1.7.4
